backend/requirements-dev.txt
# Development dependencies for code quality and testing

# Code formatting and linting
black==23.12.1
isort==5.13.2
pylint==3.0.3
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-bugbear==24.12.12

# Type checking
mypy==1.8.0
types-requests==2.31.0.20240125
types-redis==4.6.0.20240106

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.2
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==22.0.0

# Security
bandit==1.7.5
safety==3.0.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development tools
pre-commit==3.6.0
tox==4.11.4
watchdog==3.0.0

# API testing
httpx==0.26.0
requests-mock==1.11.0

# Database testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Environment management
python-dotenv==1.0.0